tests/requirements.txt
behave==1.2.5
docker-py==1.8.1
docker-compose
